+++ website/src/react-native/css/react-native.css
@@ -1338,6 +1338,30 @@ div[data-twttr-id] iframe {
   display: none;
 }
 
+
+/** Showcase **/
+
+.home-showcase-section {
+  max-width: 800px;
+  margin: 20px auto 40px auto;
+  text-align: center;
+}
+
+.home-showcase-section p {
+  max-width: 540px;
+  margin: 0 auto;
+}
+
+.footnote {
+  font-size: 12px;
+  color: rgba(0, 0, 0, 0.4);
+}
+
+.home-showcase-section .showcase img {
+  width: 110px;
+  border-radius: 20px;
+}
+
 .showcaseHeader {
   padding-bottom: 15px;
   padding-top: 15px;
@@ -1345,7 +1369,7 @@ div[data-twttr-id] iframe {
 }
 
 .showcase {
-  margin: 30px auto 10px auto;
+  margin: 30px auto 30px auto;
   width: 100%;
   display: inline-block;
   text-align: center;
@@ -1390,6 +1414,11 @@ div[data-twttr-id] iframe {
 
 .showcase img {
   width: 100px;
+  border-radius: 20px;
+}
+
+.pinned img {
+  width: 150px;
   border-radius: 20px;
 }
